==> Deep Learning/1. Data Preprocessing.ipynb <==
import pandas as pd
import numpy as np
from tqdm import tqdm
from os import listdir
from PIL import Image

df = pd.read_csv("CSV/MovieGenre_final.csv", encoding='ISO-8859-1')
df.shape
df.head()

genre_list = df['Genre']
print(len(genre_list))

def find_genres(genre):
  start = 0
  set_of_genre = []
  for i in range(len(genre)):
    k = 0
    substring = ""
    if (genre[i] == '|'):
      substring = genre[start:i]
      start = i + 1
      k = 1
    if (i == len(genre) - 1):
      substring = genre[start:i+1]
      k = 1
    if (k == 1):
      set_of_genre.append(substring)
  return (set_of_genre)

all_genre = []
Genre_list = df['Genre']

for i in range (len(Genre_list)):
    
    set_of_genre = find_genres(Genre_list[i])
    
    for j in range (len(set_of_genre)):
        all_genre.append(set_of_genre[j])
        
uniq, counts = np.unique(all_genre, return_counts=True)
print("Number of unique genres:", len(uniq))
print("Unique genres:", uniq)
dict(zip(uniq, counts))

def multi_hot_encoded_labels(img_id, genre):
    col_names = [
        'Action', 'Adult', 'Adventure', 'Animation', 'Biography', 'Comedy', 'Crime',
        'Documentary', 'Drama', 'Family', 'Fantasy', 'Film-Noir', 'Game-Show', 'History',
        'Horror', 'Music', 'Musical', 'Mystery', 'News', 'Reality-TV', 'Romance',
        'Sci-Fi', 'Short', 'Sport', 'Talk-Show', 'Thriller', 'War', 'Western'
    ]
    
    set_of_genre = find_genres(genre)
    
    row = []
    row.append(img_id)
    
    for i in range(len(col_names)):
        found = 0
        for j in range (len(set_of_genre)):
          if (set_of_genre[j]==col_names[i]):
            found = 1
            break
        row.append(found)
    row.append(genre)
    return row  

df = pd.read_csv("CSV/MovieGenre_final.csv", encoding="ISO-8859-1")
all_data = []

for index, row in tqdm(df.iterrows()):
    path = row['Image_Paths']
    genre = row['Genre']
    row = multi_hot_encoded_labels(path, genre)
    
    all_data.append(row)

col_names = [
    'Action', 'Adult', 'Adventure', 'Animation', 'Biography', 'Comedy', 'Crime',
    'Documentary', 'Drama', 'Family', 'Fantasy', 'Film-Noir', 'Game-Show', 'History',
    'Horror', 'Music', 'Musical', 'Mystery', 'News', 'Reality-TV', 'Romance',
    'Sci-Fi', 'Short', 'Sport', 'Talk-Show', 'Thriller', 'War', 'Western'
]

np.savetxt("CSV/Multi_hot_encoded_data.csv", np.asarray(all_data), fmt='%s', delimiter=" ")   
np.savetxt("CSV/Encoded_data_column_lookup.csv", np.asarray(col_names), fmt='%s', delimiter=" ")

df_encoded = pd.read_csv("CSV/Multi_hot_encoded_data.csv", delimiter=" ", 
                  names =  ['Img-paths', 'Action', 'Adult', 'Adventure', 'Animation', 'Biography', 'Comedy', 'Crime',
                  'Documentary', 'Drama', 'Family', 'Fantasy', 'Film-Noir', 'Game-Show', 'History',
                  'Horror', 'Music', 'Musical', 'Mystery', 'News', 'Reality-TV', 'Romance',
                  'Sci-Fi', 'Short', 'Sport', 'Talk-Show', 'Thriller', 'War', 'Western', 'Genre'])
df_encoded.head()

# Train 80 / Validation 15 / Test 5

df = pd.read_csv("CSV/Multi_hot_encoded_data.csv", delimiter=" ")
random_seed = 50
train_df = df.sample(frac=0.70, random_state=random_seed)
tmp_df = df.drop(train_df.index)
test_df = tmp_df.sample(frac=0.1, random_state=random_seed)
valid_df = tmp_df.drop(test_df.index)

print("Train_df:",len(train_df))
print("Val_df:",len(valid_df))
print("Test_df:",len(test_df))

np.savetxt("CSV/Train.csv", train_df, fmt='%s', delimiter=" ")
np.savetxt("CSV/Test.csv", test_df, fmt='%s', delimiter=" ")
np.savetxt("CSV/Valid.csv", valid_df, fmt='%s', delimiter=" ")
