==> Deep Learning/2. Custom Model.ipynb <==
import tensorflow as tf
from keras.models import Sequential
from keras.layers import Dense, Dropout, Flatten
from keras.layers import Conv2D, MaxPooling2D
import keras.utils as image
from keras.callbacks import EarlyStopping
from keras.callbacks import ModelCheckpoint
from keras.preprocessing.image import ImageDataGenerator
import pandas as pd 
import numpy as np 
from tqdm import tqdm
import matplotlib.pyplot as plt
from keras import models
from keras import layers
from keras import optimizers
from PIL import Image
from sklearn.model_selection import train_test_split
from statistics import mean

def arrange_data(df):
    image_data = []
    img_paths = np.asarray(df.iloc[:, 0])

    for i in tqdm(range(len(img_paths))):          
        img = image.load_img(img_paths[i],target_size=(200,150,3))
        img = image.img_to_array(img)
        img = img/255
        image_data.append(img)
    
    X = np.array(image_data)
    Y = np.array(df.iloc[:,1:29])
    
    print("Shape of images:", X.shape)
    print("Shape of labels:", Y.shape)
    
    return X, Y

print("Processing Train Dataset...")
train_df = pd.read_csv("CSV/Train.csv", delimiter=" ")
X_train, Y_train = arrange_data(train_df)

print("Processing Validation Dataset...")
val_df = pd.read_csv("CSV/Valid.csv", delimiter=" ")
X_val, Y_val = arrange_data(val_df)

num_classes = 28  

model = Sequential()

model.add(Conv2D(filters=16, kernel_size=(5, 5), activation="relu", input_shape=(200, 150, 3)))
model.add(MaxPooling2D(pool_size=(2, 2)))
model.add(Dropout(0.25))

model.add(Conv2D(filters=32, kernel_size=(5, 5), activation="relu"))
model.add(MaxPooling2D(pool_size=(2, 2)))
model.add(Dropout(0.25))

model.add(Conv2D(filters=64, kernel_size=(5, 5), activation="relu"))
model.add(MaxPooling2D(pool_size=(2, 2)))
#model.add(Dropout(0.25))

model.add(Conv2D(filters=64, kernel_size=(5, 5), activation="relu"))
model.add(MaxPooling2D(pool_size=(2, 2)))
#model.add(Dropout(0.25))

model.add(Flatten())
model.add(Dense(128, activation='relu'))
model.add(Dropout(0.5))
model.add(Dense(64, activation='relu'))
model.add(Dropout(0.5))
model.add(Dense(num_classes, activation='sigmoid'))

model.summary()

import tensorflow as tf

with tf.device('/cpu:0'):
    print("CPU working")
    model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['accuracy'])
    history = model.fit(X_train, Y_train, epochs=150, validation_data=(X_val, Y_val), batch_size=64)
    model.save('Model_Custom.h5')
