==> Deep Learning/4. Test.ipynb <==
import keras
import pandas as pd 
import numpy as np 
from tqdm import tqdm
import keras.utils as image
import matplotlib.pyplot as plt
import cv2
from PIL import Image
from keras.models import load_model

def arrange_data(df):
    image_data = []
    img_paths = np.asarray(df.iloc[:, 0])
    
    for i in tqdm(range(len(img_paths))):
        img = image.load_img(img_paths[i],target_size=(200,150,3))
        img = image.img_to_array(img)
        img = img/255
        image_data.append(img) 
        
    X = np.array(image_data)
    Y = np.array(df.iloc[:,1:29])
    
    return X, Y

def accuracy_score(test_path, model_path):
    test_df = pd.read_csv(test_path, delimiter=" ")
    X_test, Y_test = arrange_data(test_df)

    model = load_model(model_path)
    pred = model.predict(np.array(X_test))

    count = 0
    for i in tqdm(range(len(pred))):
        value = 0
        
        first3_index = np.argsort(pred[i])[-3:]
        correct = np.where(Y_test[i] == 1)[0]
        
        for j in first3_index:
            if j in correct:
                value += 1
                
        if (value > 0):
            count = count + 1
    
    print("Total Test Dataset:", len(pred))
    print("Correctly Predicted:", count)
    print("Accuracy:", count/len(pred))

accuracy_score("CSV/Test.csv", "Model_Custom.h5")

accuracy_score("CSV/Test.csv", "Model_VGG.h5")
